--- TestCase.php
@@ -1,16 +1,0 @@
-<?php
-
-class TestCase {
-    public $name;
-    public function __construct($name)
-    {
-        $this->name = $name;
-    }
-
-    public function run()
-    {
-        $method = $this->name;
-
-        $this->$method();
-    }
-}
